Commit Message: Initial Version - PassGEN 1.0
Description:
This commit introduces the first version of the password generator script to the repository.
The script is designed to create secure and random passwords to enhance user security in various applications and systems.

Features:
- Generates random passwords with customizable length.
- Includes a variety of characters, including uppercase letters, lowercase letters, numbers, and special symbols.
- Added a function to calculate the total number of possible combinations of password that the script can generate with the specified length.
- Include an encrypted personal storage (multi-user) to store passwords.

Usage:
Users can run the script from the command line to generate strong passwords and store them in a secure environment.

Future Improvements:
- Integration with a GUI.
- Integration with web applications.

Author: Nicola Staniscia
Creation Date: 2023-12-23
Version: 1.0